% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_cities.R
\name{list_cities}
\alias{list_cities}
\title{List Available Cities}
\usage{
list_cities()
}
\value{
A tibble with columns \code{city_id} and \code{city_name}.
}
\description{
Retrieves a list of cities available for filtering air quality data.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get all available cities
cities <- list_cities()
head(cities)
\dontshow{\}) # examplesIf}
}
