% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_parameters.R
\name{list_parameters}
\alias{list_parameters}
\title{List Available Measurement Parameters}
\usage{
list_parameters()
}
\value{
A tibble with details on available parameters.
}
\description{
Retrieves a list of measurement parameters available for querying.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get all available parameters
params <- list_parameters()
print(params)
\dontshow{\}) # examplesIf}
}
