% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_stations.R
\name{list_stations}
\alias{list_stations}
\title{List Available Stations}
\usage{
list_stations()
}
\value{
A tibble of station metadata.
}
\description{
Retrieves a metadata table of available stations from the air quality service.
}
\examples{
\donttest{
# Get all available stations
stations <- list_stations()
head(stations)

# Filter stations by city
library(dplyr)
ankara_stations <- stations |>
  filter(grepl("Ankara", city_name, ignore.case = TRUE))
}
}
